requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0
huggingface_hub
opencv-python>=4.6.0
numpy>=1.23.0
matplotlib>=3.6.0
pyyaml>=5.1
scikit-learn>=1.0.0
pybind11>=2.10.0
lap>=0.4.0
